Actually test for test_rmm flag

+++ Jenkinsfile
@@ -360,10 +360,12 @@ def TestPythonGPU(args) {
     def docker_args = "--build-arg CUDA_VERSION=${args.host_cuda_version}"
     def mgpu_indicator = (args.multi_gpu) ? 'mgpu' : 'gpu'
     sh "${dockerRun} ${container_type} ${docker_binary} ${docker_args} tests/ci_build/test_python.sh ${mgpu_indicator}"
-    sh "rm -rfv build/ python-package/dist/"
-    unstash name: "xgboost_whl_rmm_cuda${artifact_cuda_version}"
-    unstash name: "xgboost_cpp_tests_rmm_cuda${artifact_cuda_version}"
-    sh "${dockerRun} ${container_type} ${docker_binary} ${docker_args} tests/ci_build/test_python.sh ${mgpu_indicator}"
+    if (args.test_rmm) {
+      sh "rm -rfv build/ python-package/dist/"
+      unstash name: "xgboost_whl_rmm_cuda${artifact_cuda_version}"
+      unstash name: "xgboost_cpp_tests_rmm_cuda${artifact_cuda_version}"
+      sh "${dockerRun} ${container_type} ${docker_binary} ${docker_args} tests/ci_build/test_python.sh ${mgpu_indicator}"
+    }
     deleteDir()
   }
 }
